Add comments to state
